// frontend/pizzaiolo/src/app/model/pizzaiolo/ingredientesEditables.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Version editable de los ingredientes.
 */
export interface IngredientesEditables {
    /**
     * Identificador del ingrediente.
     */
    idIngredient: number;
    /**
     * Nombre del ingrediente.
     */
    name?: string;
    /**
     * Precio del ingrediente por unidad.
     */
    price?: number;
    /**
     * Clasificación por tipo de ingrediente.
     */
    type?: IngredientesEditables.TypeEnum;
}
export namespace IngredientesEditables {
    export type TypeEnum = 'base' | 'otros' | 'salsa';
    export const TypeEnum = {
        Base: 'base' as TypeEnum,
        Otros: 'otros' as TypeEnum,
        Salsa: 'salsa' as TypeEnum
    };
}